faktura id Int to Long change

+++ app/src/main/java/com/example/photoapp/features/faktura/data/FakturaDao.kt
@@ -18,7 +18,7 @@ interface FakturaDao {
     fun getAllFaktury(): List<Faktura>
 
     @Query("SELECT * FROM Faktura WHERE id = :id")
-    fun getById(id: Int): Faktura?
+    fun getById(id: Long): Faktura?
 
     @Insert
     fun insert(faktura: Faktura): Long
